Redaction System/requirements.txt
streamlit==1.38.0
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
spacy==3.7.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl
python-levenshtein==0.25.1
pandas==2.2.2